refactor: SwaggerConfig 로 이름 변경 및 수정
