--- src/main/java/com/followme/userserver/config/OpenApiConfig.java
@@ -1,18 +1,0 @@
-package com.followme.userserver.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class OpenApiConfig {
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-            .info(new Info()
-                .title("User Service API")
-                .version("v1.0.0")
-                .description("사용자 및 배송 담당자 관리 API"));
-    }
-}
